Enable targets to be displayed without changing the hash (URL)
Display target manually when there is an invalid URL by calling the targets through the router and displaying. Variable added and passed to router when navigating from home to preserve data that signifies where user just navigated to; this is important information to keep to prevent the user from getting stuck on the notfound page and to keep the back button functional. NotFound.controller.js() logic changed to reflect these changes